Stacks the condition for Safari support and Chrome support all in to one
This puts the IF condition for detecting Chrome PiP support into the same line as the one that sets it up for Safari. I tested it in both browsers, as well as Firefox. In browsers where PiP is supported, they each respond to their own call and not the other (Safari only responds to the Webkit check, Chrome only responds to the document check). Firefox currently supports neither standard, so there's no issue there.