More accurate slide bar control on video playback
Hello. I notice that it is difficult to do a precise time jump if the slider icon displays right under the cursor, which often happens when you just want to skip or go back a few seconds.
It seems that the slider icon covers the control bar so that the cursor clicks on the icon, not the bar. I try to fix it by scaling down the size of the slider to 0 while hovering and it works, though it may not look fancy.
I have checked some streaming sites like youtube and twitch and found that they did not have such problem. But I failed to figure out how they make it, as I am really new to web coding. It would be much better if someone else can fix it elegantly.