Remove blurhash from listView
As-is, correctly implementing blurhash on the listView would lead to significant changes to listView itself, to cardBuilder and to various places calling listView.
As a hotfix for 10.6, it is less invasive to revert the changes adding blurhash to listView and to postpone adding it to that component in 10.7, by removing the custom image creationg in order to use a card.