Make mediaInfo report viewing progress
On the info page for video files, if the user has partially viewed
the video, the number of minutes remaining will be displayed along
with the total number of minutes in the file. Also, the 'Ends at'
time will be calculated from the resume point, rather than the
beginning of the video.