Change getDeviceIcon to get icon from a list.
The previous getDeviceIcon function was just a massive switch statement
which was not only painful to look at but also means that we match using
equals instead of contains. "PeterGriffin-Firefox" is still a Firefox
but was not matched because it does not equals "Firefox".

This change updates the function to get the icon from a list by matching
the appName/deviceName with names from a list to get an icon.