Fix CI package building
Build packages as a non-root user.

Build pckages on a pull request to ensure packaging regressions get
identified before landing.

Reduce package installation to only what is necessary to build the
source.

Use more natural booleans in shell scripts.

Add datestamp to the Release: for unstable builds instead of mangling
up the version.

Move ARG processing in Dockerfiles to the end of the file so that a
simple change in an ARG value does not cause O/S installation/upgrade
layers to be repeated.

Signed-off-by: Brian J. Murrell <brian@interlinx.bc.ca>