Fix theme song playback when switching items with different themes
When you change an item with a theme to an item with a different theme, the
backdrop tries to access the player. The latter has not yet started playing the
new theme (`streamInfo` is null), but is already considered playing. A runtime
error occurs when referring to null `streamInfo` (to check if the content type
matches).

To fix the problem, set `streamInfo` just before playback starts.