Fix old streamInfo on stream change
When the client changes the stream (fail to play) `streamInfo`
contains the old play session and the next progress report
will update the wrong session. The server then responds with
the wrong playback method until the next progress report.

Update `streamInfo` before playback start so that the data
is up to date.