Make book table of contents prettier
Changes the way the table of contents looks. The two biggest parts are hiding the bullets in the list, making each link black and hiding the link underline. This makes them look more like buttons than web links. The #dialogToc changes resize the table of contents based on the length of the list and the length of the longest chapter name rather than have the table of contents take up most of the screen for a short and small list. The link text size was also increased from 92% to 120%.