Add epubjs theme support to the book player
Add a dark theme to the book player via epubjs. The theme itself uses
the dark theme of Jellyfin and only changes the colour of the text.

For everything but the text, the dark theme from Jellyfin is retained.
Should any other theme than the dark theme be chosen by the user, epubjs
falls back to it's default theme.