Autoset audio/subtitles does not affect transcode
Addresses jellyfin-web#3994

The autoset options only take effect on non-transcoded media-- the
transcode URL is never updated! This means that if the selected
audio does not match the default, the audio/subs are in the default
but the UI reflects the auto-set variant (showing the wrong track).

When doing the auto-set logic, also trivially change the audio index in
the transcode URL to match.

This fixes it in most cases, but might have some edge cases that need
to go through the full validation to recreate the URL. I don't know the
code well enough for that, though.