remove dependency on `@uupaa/dynamic-import-polyfill`
Removed the dependency on `@uupaa/dynamic-import-polyfill` and switched to using raw `import()`, because the polyfill were having issues with `async`/`await` in plugin modules which were using `await` at the top-level in the module to e.g. do dynamic imports because plugins don't have a base-point to do static imports of other modules.