Playback info: Add frame rate, bit depth, color space, color transfer and pixel format
This is an admittedly totally subjective list of stats that I find
useful and have so far had to look up in the 'Media info' instead
(which is currently barely usable on TV for me, cannot scroll down
and so only see about half of the video stream stats).

Also display the original VideoRangeType for DoVi media, not just
the 'nice' VideoDoViTitle (which e.g. currently doesn't differentiate
between HDR10 and HDR10+ fallbacks).