Change the class nextUpSection to an ID in index.html.
I decided to make this change because nothing was being assigned to the nextUpSection class, and it makes more sense for this class to be considered an ID, thus providing a greater level of customization if someone like me is creating their own theme. Another factor that led me to this change is the fact that the seasons section has an ID, but the next episodes section does not.